Configurable Bodymovin & Bodymovin_light support
This is just a pretty simple change that allows you to configure which
bodymovin library you want to use in your `webpack.config.js`. I needed
to use expressions (and thus the full bodymovin.js), and I figured
others might appreciate the option as well.

By default it will still use bodymovin_light, and doesn’t require
configuration in the webpack config, so it should be backwards
compatible, and anyone that doesn’t need / want to configure the
library don’t have to.